[Cleanup][Benchmark] Reuse audio metric definitions for PCM format
